refactor: Migrate to module execution and use faster-whisper backend
