+++ burnin_subtitle_checker/__main__.py
@@ -1,0 +1,5 @@
+import sys
+from burnin_subtitle_checker.cli import main
+
+if __name__ == "__main__":
+    sys.exit(main())
