+++ requirements.txt
@@ -1,5 +1,5 @@
 # Core dependencies
-openai-whisper>=20231117
+faster-whisper>=1.0.0
 opencv-python>=4.8.0
 pytesseract>=0.3.10
 rapidfuzz>=3.5.0
